services/user-service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
dotenv
gunicorn
psycopg2-binary